refactored scopes folder hierarchy
