+++ commands/container-registry/token/scopes/list.go
@@ -1,0 +1,72 @@
+package scopes
+
+import (
+	"context"
+	"github.com/ionos-cloud/ionosctl/commands/container-registry/registry"
+	"github.com/ionos-cloud/ionosctl/pkg/constants"
+	"github.com/ionos-cloud/ionosctl/pkg/core"
+	"github.com/ionos-cloud/ionosctl/pkg/printer"
+	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
+)
+
+func TokenScopesListCmd() *core.Command {
+	cmd := core.NewCommand(
+		context.TODO(), nil, core.CommandBuilder{
+			Namespace:  "token",
+			Resource:   "scope",
+			Verb:       "list",
+			Aliases:    []string{"l", "ls"},
+			ShortDesc:  "Get a token scopes",
+			LongDesc:   "Use this command to get a token scopes of a container registry.",
+			Example:    "ionosctl container-registry token scope list --registry-id [REGISTRY-ID], --token-id [TOKEN-ID]",
+			PreCmdRun:  PreCmdTokenScopesList,
+			CmdRun:     CmdGetTokenScopesList,
+			InitClient: true,
+		},
+	)
+
+	cmd.AddStringFlag("registry-id", "r", "", "Registry ID")
+	_ = cmd.Command.RegisterFlagCompletionFunc(
+		"registry-id", func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+			return registry.RegsIds(), cobra.ShellCompDirectiveNoFileComp
+		},
+	)
+	cmd.AddStringFlag("token-id", "t", "", "Token ID")
+	_ = cmd.Command.RegisterFlagCompletionFunc(
+		"token-id", func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+			return TokensIds(), cobra.ShellCompDirectiveNoFileComp
+		},
+	)
+
+	// TODO: add --all flag
+	// cmd.AddBoolFlag(constants.ArgAll, constants.ArgAllShort, false, "List all scopes of all tokens of a registry.")
+
+	cmd.Command.Flags().StringSlice(constants.ArgCols, nil, printer.ColsMessage(allScopeCols))
+	_ = cmd.Command.RegisterFlagCompletionFunc(
+		constants.ArgCols,
+		func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+			return allScopeCols, cobra.ShellCompDirectiveNoFileComp
+		},
+	)
+	return cmd
+}
+
+func CmdGetTokenScopesList(c *core.CommandConfig) error {
+	reg_id := viper.GetString(core.GetFlagName(c.NS, "registry-id"))
+	token_id := viper.GetString(core.GetFlagName(c.NS, "token-id"))
+	token, _, err := c.ContainerRegistryServices.Token().Get(token_id, reg_id)
+	if err != nil {
+		return err
+	}
+
+	return c.Printer.Print(getTokenScopePrint(nil, c, &token, true))
+}
+
+func PreCmdTokenScopesList(c *core.PreCommandConfig) error {
+	err := core.CheckRequiredFlags(c.Command, c.NS, "token-id", "registry-id")
+	if err != nil {
+		return err
+	}
+	return nil
+}
